Add files via upload
The files are MATLAB codes for obtaining the diffraction spots from different wrinkled surfaces. The filenames represent the sample substrate (p), thickness (t) of the thin-film deposited on the substrate and the heating temperature (T). For example, the file "p5t150T100.m" contains the code for  diffraction spots obtained from the sample with PDMS 5 substrate deposited with 150nm thin-film of Arsenic Triselenide, heated at 100 degree Celcius, and subsequently cooled down to the room temperature for wrinkle formation.